requirements.txt
ultralytics>=8.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pillow>=8.0.0
pyyaml>=5.4.1
google-cloud-storage>=2.0.0
google-cloud-aiplatform>=1.0.0
python-dotenv>=0.19.0 